feat(network): introduce req/resp messaging
Add request-response messaging as a core feature to the network stack.

Featuring a fluent interface for request handling: `network.on(<pattern>).into_stream()` registers handlers based on matching logic, and the `respond_with_concurrent` helper makes it simple to process multiple requests in parallel without manual task orchestration.

Along with that we introduce serializable `Request` and `Response` message enums to model the communication between nodes in the network.

Under the hood, we also introduce a `cbor_codec` module that leverages the `ciborium` family of crates to perform CBOR serialization and deserialization for all request and response messages. The new `Codec<Req, Resp>` implementation conforms to libp2p’s RequestResponse interface, ensuring that our RPC traffic is both compact and efficient.

Co-authored-by: Jan Schlicht <jan@katulu.io>