feat: prototype task communication
Schedulers announce tasks to workers. Available workers respond and once
enough workers have responded, the task will be distributed among the
participating workers. The necessary messages and application logic is
implemented here as a prototype. Additional information needs to be
added to the messages (e.g. available resources) and the scheduler logic
needs to be configurable, it's hard-coded currently.