ci: validate Python files
Adds CI checks for Python code. Currently our Python driver is closely
tied to functionality in the Rust code. Therefore, we provide this code
as part of the Hypha project. But we will treat each driver as their own
project, with their own dependencies. As a consequence, dependencies are
handled per-project and the CI scripts will run the verification tools
separately for each project.