feat: status message communication
This PR introduces proper `JobStatus` communication between Worker and Scheduler.
Currently, we utilize AIM to visualize training progress with a local AIM server
running at the scheduler.

The communication between the Scheduler and AIM is handled with an additional
HTTP server, since AIM depends on Python and the Scheduler is running in RUST.
The communication between RUST and Python is done with JSON. This includes the
communication Job <-> Bridge on the Worker and StatusBrider <-> AIM on the
Scheduler.

This builds the foundation to utilize status messages, thus they aren't all used
at the moment.