fix: increase stream accept limit
When accepting incoming streams of data, 'libp2p-stream' applies
backpressure by limiting the amount of stream that can be accepted
concurrently. This limit will deny incoming streams. Especially in
artificial setups like tests streams might need to be accepted
concurrently though and a stream denial will result in errors. To
resolve this, we allow for some leeway in accepting streams.