feat: react to node failures
Ensures that a failure to renew leases in the scheduler results in a
failed task. For this, node and task handling has been refactored. Task
status updates are tied to the lifetime of a task, a 'Task' instance
provides a stream of status messages. Likewise, node leases are tied to
a node instance. With these refactorings, we can handle different error
cases in a simple way.

For now, an error when trying to renew a lease will stop the scheduler,
which in turn also stops scheduled tasks on nodes. Further work is
needed to allow for actual fault-tolerance by, e.g., using the allocator
to request work from other nodes instead.