feat: batch-level scheduling
This introduces scheduling on a per batch basis which was first desribed
in #94. The change introduces a new Codec for communication status
messages between worker and scheduler. This is the main channel to
exchange training related messages. Furthermore, the Scheduler keeps
track of worker processing speed and uses this information to
determin when a local-round is finished and a DiLoCo update should
be performed.