feat: improved transformers support
This change introduces support for transformer Auto classes (Model or Processor). This allows
to support all kind of transfomers models with their different preprocessors. Since
input names to models and preprocessors vary, the change introduces an `input-names` list that
defines the input fields from the dataset for the model/preprocessor.