fix: adjust parameter server to streamline training
Change the parameter server to accept all incoming updates all the time and merge a given set upon
request instead of explicity instructing the parameter server to listen for a given set of worker
updates and them merge all updates that arrived. This removes a massive delay when wokers send their
updates caused by timing issues between instructing the parameter server to recieve and the workers
so send updates.