feat: hash-based data slice addressing
Uses hashes instead of indices to address the individual slices of a
dataset. This will make it easier for workers to keep track of slices
they already know and implement caching. When starting a data node,
it'll compute hashes of all files of a dataset and announce these in the
DHT.