refactor: handle scheduler fetches in bridge
Instead of using a separate connector for fetching data from schedulers,
we handle this directly in the bridge. These resources are used to
request data slices from schedulers, the actual data fetching is handled
separately. By moving this handler into the bridge, this allows us to
decide if we actually want to fetch data from peers based on the
response from schedulers.