feat: simple data slice cache
Cache data slices in the simplest way possible: We store data slices in
files. If we already have a file for a slice, which is uniquely
identifiable by its hash, we don't download that slice again. For the
moment, this works fine, because we only clean up created files when
stopping a worker.