feat: support multiple data providers
Handle the case of multiple data providers providing the same dataset.
For that case we assume that a dataset name is used uniquely for the
same dataset. With that assumption, we can inform workers of the
multiple data providers for a dataset and have them choose between them
when downloading data slices. Right now, workers choose randomly if
there are multiple data providers available.