feat(worker): revoke leases on failed tasks
Move executors to a unified Status enum, so that task outcomes are observable. The arbiter uses     these statuses to discard leases when task (aka jobs) fail. This results in the scheduler re-allocating a worker for that tasks and then re-dispatching the it. This way a failure in an executor does not fail the training job. It assumes that there is just one task per lease.