feat: use request/response for data records
Data providers announce their available dataset using the DHT.
Schedulers need a list of data slices of a dataset (= data record) if
that dataset will be used in a job. We used to store this data record in
the DHT as well. But DHT values are restricted in size and a data record
can become quite large for some datasets. This commit changes the
current behavior to no longer store data records in DHT. Instead, a data
provider provides a request/response API for schedulers to request a
data record for a dataset. By using request/response, we're no longer
limited in the size a data record can have.